# api-design/API.yaml
swagger: '2.0'
info:
  title: Religram API
  description: Tài liệu đặc tả API cho phía fontend.
  version: 1.0.0
host: religram.relipa-test.online
basePath: /api/v1
...
schemes:
  - https
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Login
      description: Nhập username (email) và password để login vào ứng dụng
      operationId: 'login'
      parameters:
        - name: body
          in: body
          required: true
          description: Thông tin đăng nhập gồm user (email) và password
          schema:
            $ref: '#/definitions/LoginRequest'
      produces:
        - application/json
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/LoginResponse'
        401:
          description: Unauthorized. Username/password incorrected.
  /auth/signup:
    post:
      tags:
        - auth
      summary: Signup new user
      description: Nhập các thông tin user để đăng ký tài khoản mới
      parameters:
        - name: body
          in: body
          required: true
          description: Thông tin đăng ký tài khoản
          schema:
            $ref: '#/definitions/SignupRequest'
      operationId: 'signup'
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/LoginResponse'
        400:
          $ref: '#/responses/ValidationError'
        409:
          $ref: '#/responses/UserAlreadyExistError'
...
  /auth/changepassword:
    put:
      tags:
        - auth
      summary: Change password
      operationId: 'changePassword'
      parameters:
        - name: body
          in: body
          description: Đẩy thông tin current password và password mới
          schema:
            type: object
            properties:
              currentPassword:
                type: string
              newPassword:
                type: string
      responses:
        200:
          description: Successful operation
        400:
          description: Current password not match or new password invalid
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int64
              message:
                type: string
        401:
          $ref: '#/responses/UnauthorizedError'
      security:
        - religram_auth: []
...
  /user:
    put:
      tags:
        - user
      summary: Update current User info
      description: Dùng update thông tin user
      operationId: 'updateUserInfo'
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: body
          description: Trường avatar gửi lên base64 của avatar
          in: body
          schema:
            $ref: '#/definitions/UpdateUserRequest'
      responses:
        200:
          schema:
            $ref: '#/definitions/UpdatedUserResponse'
          description: Successful operation
        401:
          $ref: '#/responses/UnauthorizedError'
        409:
          $ref: '#/responses/UserAlreadyExistError'
        default:
          description: Unexpected error
      security:
        - religram_auth: []
...
responses:
  UnauthorizedError:
    description: API key is missing or invalid
    headers:
      Authorization:
        description: Bearer {token}
        type: string
  UserAlreadyExistError:
    description: Xảy ra khi đăng ký/update username đã tồn tại trên hệ thống
    schema:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
        message:
          type: string
  ValidationError:
    description: Xảy ra khi các thuộc tính gửi lên không đúng validdation.
    schema:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
        message:
          type: array
          xml:
            name: hashtags
            wrapped: true
          items:
              type: object
              properties:
                key:
                  type: string
                message:
                  type: string
securityDefinitions:
  religram_auth:
    type: "apiKey"
    in: header
    name: Authorization
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      fullname:
        type: string
      email:
        type: string
      avatar:
        type: string
    xml:
      name: User
...
  LoginRequest:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    xml:
      name: LoginRequest
  SignupRequest:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
      fullname:
        type: string
      email:
        type: string
    xml:
      name: SignupRequest
  LoginResponse:
    type: object
    properties:
      user:
        $ref: '#/definitions/User'
      token:
        type: string
    xml:
      name: LoginResponse
  UpdateUserRequest:
    type: object
    properties:
      username:
        type: string
      fullname:
        type: string
      avatar:
        type: string
        description: Base 64 của Image
    xml:
      name: UpdateUserRequest
  UpdatedUserResponse:
    type: object
    properties:
      user:
        $ref: '#/definitions/User'
      token:
        type: string
    xml:
      name: UpdatedUserResponse